move migration scripts
